dataset-aidoctor-chatbot/requirements.txt
fastapi
uvicorn
joblib
scikit-learn
numpy
google-generativeai
python-dotenv
